requirements.txt
# Jupyter
jupyterlab==4.0.9
jupyter-resource-usage
jupyter-server
dask-labextension
ipywidgets

# dask
dask[complete]
dask-ml==2023.3.24
graphviz
flox

# geospatial
numpy==1.26.4 #last release before 2.0
xarray[complete]==2024.10.0
rioxarray
pandas
geopandas
odc-geo==0.4.2
fiona

# ML
scikit-learn==1.3.2 #models trained on this version
lightgbm==4.2.0
shap

# plotting
contextily
seaborn
folium
odc-ui
mapclassify

# analysis
scikit-image
xskillscore